Add soundcloud upload support

+++ post_recording.py
@@ -1,19 +1,33 @@
 #!/usr/bin/python
+import yaml
 
-from pydub import AudioSegment
 import sys
 import getopt
 import os
+import datetime
+import soundcloud
+
+from pydub import AudioSegment
+from pytz import timezone    
+
+
 
 def main(argv):
     inputfile = ''
     outputfile = ''
     try:
-        opts, args = getopt.getopt(argv,"hi:s:")
+        opts, args = getopt.getopt(argv,"hpi:s:")
     except getopt.GetoptError:
         print 'post_recording.py -i <input_file> -s <terminating_silence>'
         sys.exit(2)
 
+    # Load in the config file
+    config = yaml.load(file("config/settings.yml"))
+
+    # Configure some defaults
+    preserve_files = False
+
+    # Load in command line options
     for opt, arg in opts:
         if opt == '-h':
             print 'post_recording.py -i <input_file> -s <terminating_silence>'
@@ -22,6 +36,8 @@ def main(argv):
             input_file = arg
         elif opt in ("-s"):
             terminating_silence = int(arg)
+        elif opt in ("-p"):
+            preserve_files = True
 
     # Create the mp3 file
     song = AudioSegment.from_wav(input_file)
@@ -32,8 +48,30 @@ def main(argv):
     basename, ext = os.path.splitext(input_file)
     output_file = basename + '.mp3'
     song.export(output_file, format="mp3")
+
+    # create client soundcloud object with app and user credentials
+    client = soundcloud.Client(client_id=config['soundcloud']['client_id'],
+            client_secret=config['soundcloud']['client_secret'],
+            username=config['soundcloud']['username'],
+            password=config['soundcloud']['password'])
+
+    # Get the date for the file name
+    local_timezone = timezone(config['soundcloud']['timezone'])
+    current_time = datetime.datetime.now(local_timezone)
+    formatted_time = current_time.strftime('%l:%M %p %B %d, %Y')
+    
+    # upload audio file
+    track = client.post('/tracks', track={
+            'title': 'PiAno Recording - ' + formatted_time,
+                'asset_data': open(output_file, 'rb')
+                })
+
+    # print track link
+    # print track.permalink_url
+
     # Clean up the wav
-    os.remove(input_file)
+    if not preserve_files:
+        os.remove(input_file)
 
 
 if __name__ == "__main__":
